added first round trained model
Unzip the models to use. Trained on the linked motor imagery dataset. The dataset has 5 classes but for efficiency and accuracy with consumer grade equipment, I reduced the number of commands to 3 and the actual pong game only uses the classes 'left fist' and 'right fist'. Accuracy goes from around 60% with all classes to 90%+ with just 3.